Revert "Guard browser helpers against invalid viewports"
This reverts commit 6181c81059e0638dee20c981ec5bde6fa1456e9d.
