+++ package/ego-browser/src/driver/observe.test.mjs
@@ -8,10 +8,6 @@ import {
 import { captureScreenshot } from "../../dist/src/driver/observe.js";
 import { setOverrides } from "../../dist/src/state.js";
 
-function runtimeValue(value) {
-  return { result: { value } };
-}
-
 function withCdpRuntime(fn) {
   const previous = globalThis.ego;
   const sent = [];
@@ -100,116 +96,4 @@ test("captureScreenshot skips page metric JavaScript while a native dialog is pe
 
   assert.equal(writes.length, 1);
   assert.equal(writes[0].path, "/tmp/ego-browser-dialog-shot.png");
-});
-
-test("captureScreenshot rejects zero viewport before calling Page.captureScreenshot", async () => {
-  const calls = [];
-  const restore = setOverrides({
-    cdpOverride(method, params, sessionId) {
-      calls.push({ method, params, sessionId });
-      if (method === "Runtime.evaluate") {
-        if (params.expression === "window.devicePixelRatio") {
-          return runtimeValue(1);
-        }
-        return runtimeValue(
-          JSON.stringify({
-            url: "https://example.com/zero",
-            title: "Zero",
-            w: 0,
-            h: 0,
-            sx: 0,
-            sy: 0,
-            pw: 320,
-            ph: 180,
-          }),
-        );
-      }
-      return {};
-    },
-  });
-  try {
-    await assert.rejects(
-      () => captureScreenshot("/tmp/ego-browser-zero-shot.png"),
-      (error) => {
-        assert.equal(error.code, "EGO_INVALID_VIEWPORT");
-        assert.match(error.message, /captureScreenshot/);
-        assert.match(error.message, /viewport=0x0/);
-        return true;
-      },
-    );
-  } finally {
-    restore();
-  }
-
-  assert.equal(
-    calls.some((call) => call.method === "Page.captureScreenshot"),
-    false,
-  );
-});
-
-test("captureScreenshot rejects explicit clips on zero viewport unless raw", async () => {
-  const calls = [];
-  const writes = [];
-  const restore = setOverrides({
-    async writeFile(path, data) {
-      writes.push({ path, data });
-    },
-    cdpOverride(method, params, sessionId) {
-      calls.push({ method, params, sessionId });
-      if (method === "Runtime.evaluate") {
-        if (params.expression === "window.devicePixelRatio") {
-          return runtimeValue(1);
-        }
-        return runtimeValue(
-          JSON.stringify({
-            url: "https://example.com/zero",
-            title: "Zero",
-            w: 0,
-            h: 0,
-            sx: 0,
-            sy: 0,
-            pw: 320,
-            ph: 180,
-          }),
-        );
-      }
-      if (method === "Page.captureScreenshot") {
-        return { data: Buffer.from("png").toString("base64") };
-      }
-      return {};
-    },
-  });
-  try {
-    await assert.rejects(
-      () =>
-        captureScreenshot("/tmp/ego-browser-zero-clip.png", {
-          clip: { x: 0, y: 0, width: 100, height: 100 },
-        }),
-      /EGO_INVALID_VIEWPORT/,
-    );
-
-    assert.equal(
-      calls.some((call) => call.method === "Page.captureScreenshot"),
-      false,
-    );
-
-    await captureScreenshot("/tmp/ego-browser-zero-raw-clip.png", {
-      raw: true,
-      clip: { x: 0, y: 0, width: 100, height: 100 },
-    });
-  } finally {
-    restore();
-  }
-
-  const rawScreenshot = calls.find(
-    (call) => call.method === "Page.captureScreenshot",
-  );
-  assert.deepEqual(rawScreenshot?.params.clip, {
-    x: 0,
-    y: 0,
-    width: 100,
-    height: 100,
-  });
-  assert.equal(writes.length, 1);
-  assert.equal(writes[0].path, "/tmp/ego-browser-zero-raw-clip.png");
 });
