--- package/ego-browser/src/driver/viewport.ts
@@ -1,99 +1,0 @@
-import {
-  ensureSession,
-  isBrowserRuntime,
-  pendingDialog,
-} from "../browser-runtime.js";
-import { js } from "../cdp-eval.js";
-import { state } from "../state.js";
-
-export type PageInfo =
-  | {
-      url: string;
-      title: string;
-      w: number;
-      h: number;
-      sx: number;
-      sy: number;
-      pw: number;
-      ph: number;
-    }
-  | { dialog: object };
-
-type WaitForValidViewportOptions = {
-  timeoutMs?: number;
-  intervalMs?: number;
-};
-
-const DEFAULT_VIEWPORT_READY_TIMEOUT_MS = 5000;
-const VIEWPORT_READY_INTERVAL_MS = 100;
-
-export async function readPageInfo(): Promise<PageInfo> {
-  if (isBrowserRuntime()) {
-    await ensureSession();
-    const dialog = pendingDialog();
-    if (dialog) {
-      return { dialog };
-    }
-  }
-  const expression =
-    "JSON.stringify({url:location.href,title:document.title,w:innerWidth,h:innerHeight,sx:scrollX,sy:scrollY,pw:document.documentElement.scrollWidth,ph:document.documentElement.scrollHeight})";
-  return JSON.parse(await js(expression));
-}
-
-export function hasDialog(info: PageInfo): info is { dialog: object } {
-  return Object.hasOwn(info, "dialog");
-}
-
-export function hasValidViewport(info: PageInfo) {
-  return !hasDialog(info) && info.w > 0 && info.h > 0;
-}
-
-export function invalidViewportError(info: PageInfo, context: string) {
-  const details = hasDialog(info)
-    ? "a native JavaScript dialog is pending"
-    : `viewport=${info.w}x${info.h}, page=${info.pw}x${info.ph}, url=${info.url}`;
-  const error: any = new Error(
-    `EGO_INVALID_VIEWPORT: ${context} requires a non-zero viewport; ${details}`,
-  );
-  error.code = "EGO_INVALID_VIEWPORT";
-  error.details = { context, pageInfo: info };
-  return error;
-}
-
-export function assertValidViewport(info: PageInfo, context: string) {
-  if (!hasValidViewport(info)) {
-    throw invalidViewportError(info, context);
-  }
-}
-
-export async function waitForValidViewport(
-  context: string,
-  options: WaitForValidViewportOptions = {},
-) {
-  const timeoutMs = options.timeoutMs ?? DEFAULT_VIEWPORT_READY_TIMEOUT_MS;
-  const intervalMs = options.intervalMs ?? VIEWPORT_READY_INTERVAL_MS;
-  const deadline = state.now() + timeoutMs;
-  let lastInfo: PageInfo | null = null;
-
-  while (state.now() <= deadline) {
-    lastInfo = await readPageInfo();
-    if (hasDialog(lastInfo) || hasValidViewport(lastInfo)) {
-      return lastInfo;
-    }
-    await state.sleep(intervalMs);
-  }
-
-  throw invalidViewportError(
-    lastInfo ?? {
-      url: "",
-      title: "",
-      w: 0,
-      h: 0,
-      sx: 0,
-      sy: 0,
-      pw: 0,
-      ph: 0,
-    },
-    context,
-  );
-}
